feat: #661 Remove `saas ci create-credentials` command, introduce `saas ci get-artifacts-bucket` command instead
